Adding a map which can be sealed/unsealed for writes on demand
The new map is used in RelMetadataQuery which contains an unbound map to which
entries are added everytime a new RelNode is created on the cluster.

RelMetadataQuery is used in RelOptCluster. For RelOptClusters that are long
lived, this can result in memory leaks when RelNodes are added to the map and
that is the only remaining reference to the RelNode preventing garbage
collection

RelMetadataQuery is not thread-safe so its map cannot be written to if it is
used across multiple threads. Instead, its map should be sealed before using
in it in a multi threaded environment.

There is no change from current behavior if the seal/unseal functionality is
not used