Relocate shaded artifacts into $internal
Previously, when working in an IDE with a downstream project, the IDE
might have hard time knowing that
`com.google.common.collect.ImmutableList` should be used and not
`com.linkedin.coral.com.google.common.collect.ImmutableList`. By
relocating to an `$internal` package, it becomes easy to configure an
IDE so that it does not offer to import shaded classes.