Introduce shaded-all artifact with dependencies actually shaded
Previously the `shaded` artifact would have some dependencies shaded and
some not. All should be equally shaded, otherwise the shaded artifact
is hard to use. Instead of fixing the `shaded` artifact, a `shaded-all`
is introduced, because project maintainers were concerned about breaking
some projects that may be depending on the factually incorrect shading
in the `shaded` one.